requirements.txt
streamlit
supabase
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
sentence-transformers
torch
requests